Smearing Tool moved to SimG4Fast
