+++ Sim/SimG4Fast/src/components/SimpleSmear.cpp
